tests/wpt/web-platform-tests/websockets/constructor/005.html

<!doctype html>
<title>WebSockets: return value</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<div id=log></div>
<script>
test(function() {
  assert_true(new WebSocket("ws://"+location.host+"/") instanceof WebSocket);
});
</script>
